refactor(ChessProgram): Service 이름 변경
