+++ src/main/java/chess/controller/ChessController.java
@@ -1,7 +1,7 @@
 package chess.controller;
 
-import chess.domain.board.Board;
 import chess.domain.ChessGame;
+import chess.domain.board.Board;
 import chess.domain.command.Command;
 import chess.domain.position.Position;
 import chess.exception.ChessException;
