+++ src/main/java/chess/domain/piece/PieceFactory.java
@@ -1,7 +1,6 @@
 package chess.domain.piece;
 
 import chess.domain.Side;
-import chess.domain.piece.*;
 import chess.exception.ChessException;
 
 public class PieceFactory {
